+++ mercury/mercury-library/agent-didcommx/src/main/scala/io/iohk/atala/mercury/model/Conversions.scala
@@ -38,6 +38,9 @@ given Conversion[Message, org.didcommx.didcomm.message.Message] with {
     msg.from.foreach(did => aux.from(did.value))
     msg.to.foreach(did => aux.to(Seq(did.value).asJava))
 
+    msg.pleaseAck.foreach { seq => // https://identity.foundation/didcomm-messaging/spec/#acks
+      aux.pleaseAck(true) // FIXME lib limitation the field pleaseAck MUST be a Array of string
+    }
     msg.ack.flatMap(_.headOption).foreach(str => aux.ack(str)) // NOTE: headOption becuase DidCommx only support one ack
     msg.thid.foreach(str => aux.thid(str))
     msg.pthid.foreach(str => aux.pthid(str))