+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/model/error/PresentationError.scala
@@ -11,4 +11,7 @@ object PresentationError {
   final case class InvalidFlowStateError(msg: String) extends PresentationError
   final case class UnexpectedError(msg: String) extends PresentationError
   final case class IssuedCredentialNotFoundError(cause: Throwable) extends PresentationError
+  final case class PresentationDecodingError(cause: Throwable) extends PresentationError
+  final case class PresentationNotFoundError(cause: Throwable) extends PresentationError
+  final case class HolderBindingError(msg: String) extends PresentationError
 }