+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/Main.scala
@@ -12,12 +12,9 @@ import io.iohk.atala.agent.walletapi.service.ManagedDIDService
 import io.iohk.atala.resolvers.DIDResolver
 import io.iohk.atala.agent.server.http.ZioHttpClient
 import org.flywaydb.core.extensibility.AppliedMigration
-import io.iohk.atala.pollux.service.{
-  JdbcSchemaRegistryService,
-  SchemaRegistryServiceInMemory,
-  VerificationPolicyServiceInMemory
-}
+import io.iohk.atala.pollux.service.{JdbcSchemaRegistryService, SchemaRegistryServiceInMemory}
 import io.iohk.atala.agent.walletapi.sql.JdbcDIDSecretStorage
+import io.iohk.atala.pollux.schema.controller.VerificationPolicyControllerInMemory
 
 object Main extends ZIOAppDefault {
 
@@ -37,7 +34,7 @@ object Main extends ZIOAppDefault {
   } yield ()
 
   def appComponents(didCommServicePort: Int, restServicePort: Int) = for {
-    _ <- Modules.didCommExchangesJob.debug.fork
+    _ <- Modules.issueCredentialDidCommExchangesJob.debug.fork
     _ <- Modules.presentProofExchangeJob.debug.fork
     _ <- Modules.connectDidCommExchangesJob.debug.fork
     _ <- Modules.didCommServiceEndpoint(didCommServicePort).debug.fork
@@ -101,10 +98,8 @@ object Main extends ZIOAppDefault {
         SystemModule.actorSystemLayer,
         HttpModule.layers,
         RepoModule.credentialSchemaServiceLayer,
-        VerificationPolicyServiceInMemory.layer,
         AppModule.manageDIDServiceLayer,
-        JdbcDIDSecretStorage.layer,
-        RepoModule.agentTransactorLayer
+        RepoModule.verificationPolicyServiceLayer
       )
     } yield app